backend/requirements.txt
# Core
Flask==3.1.0
Flask-CORS==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Limiter==3.12
python-dotenv==1.1.0
gunicorn==23.0.0

# Auth
bcrypt==4.3.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# AI / ML (install separately if needed)
# sentence-transformers==3.4.1
# faiss-cpu==1.9.0
openai==1.68.2

# Async Tasks (optional)
celery==5.4.0
redis==5.2.1

# Security
marshmallow==3.26.1

# Utilities
# numpy fails to install on latest python without GCC 8
# numpy==2.2.4
# rank_bm25==0.2.2
